Move screenshot logic to a dedicated class
